move embedded struct example into godoc
Examples are nicer when they're shown in the godoc.
